+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
 ## 0.2.11
+
+### Features
+* `Autocomplete` prop `popperProps` was renamed to `dropdownProps`
+* Added max-width css property to `Autocomplete`'s popper
+* `Select` prop `dropdownPopperProps` was renamed to `dropdownProps`
+* `Select` and `Autocomplete` css style `width: 100%` was removed (in cause of popper displaying bug)
+* Added `Dropdown`'s prop `useReferenceWidth` to make dropdown's width same as its reference
+* `Dropdown`'s prop `appendToBody` was set to `true`
+
+## 0.2.11
 ### Maintenance
 * fixed clear-icon-as-suffix-icon in storybook